moving playgame, renaming to play_game
playgame is still in root, which means it won't be linted
it also doesn't follow snake_case convention for module names
which bugs me
fix those, then import and script references
