+++ .pylintrc
@@ -76,6 +76,11 @@ disable=bad-continuation,
       trailing-whitespace,
       wrong-import-order,
       no-self-use,
+      #play_game rules
+      super-init-not-called,
+      logging-format-interpolation,
+      deprecated-lambda,
+      unnecessary-lambda,
       #issues only found in tests
       redundant-unittest-assert,
       ungrouped-imports,
