+++ sample.sh
@@ -1,4 +1,5 @@
-python playgame.py "single" \
+python hanabi_ai/play_game.py \
+    "single" \
     "hanabi_ai.players.example_discarder.Discarder" \
     "hanabi_ai.players.example_discarder.Discarder" \
     --log_dir "game_logs/results.txt" \
