+++ tests/hanabi_game_tests.py
@@ -1,5 +1,5 @@
 import unittest
-from playgame import HanabiGame
+from hanabi_ai.play_game import HanabiGame
 from argparse import Namespace
 from hanabi_ai.model.hanabi_deck import HanabiVariant
 import hanabi_ai.model.hanabi_moves as moves
