+++ tests/play_game_tests.py
@@ -1,7 +1,7 @@
-import playgame
+import hanabi_ai.play_game as play_game
 import unittest
 from sets import Set
-from playgame import HanabiGame
+from hanabi_ai.play_game import HanabiGame
 from argparse import Namespace
 from hanabi_ai.model.hanabi_deck import HanabiVariant
 from hanabi_ai.players.hanabi_player import HanabiPlayer
@@ -10,7 +10,7 @@ class PlayGameparserTests(unittest.TestCase):
 
     def test_game_simple(self):
         args = ['single', 'Discarder']
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertEqual(1, len(parsed.players))
         self.assertEquals(0, parsed.variant)
         self.assertFalse(parsed.verbose)
@@ -21,45 +21,45 @@ class PlayGameparserTests(unittest.TestCase):
 
     def test_game_iterations(self):
         args = ['single', 'Discarder', '-i', '10']
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertEquals(10, parsed.iterations)
         self.assertEquals(parsed.command, 'single')
 
     def test_game_variant(self):
         args = ['single', 'Discarder', '-r', '3']
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertEquals(3, parsed.variant)
 
     def test_game_verbose(self):
         args = ['single', 'Discarder', '-v']
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertTrue(parsed.verbose)
 
     def test_game_log_dir(self):
         args = ['single', 'Discarder', '-l', 'results.txt']
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertEquals('results.txt', parsed.log_dir)
 
     def test_game_log_stderr(self):
         args = ['single', 'Discarder', '-e', 'errors.txt']
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertEquals('errors.txt', parsed.log_stderr)
 
     def test_game_seed(self):
         args = ['single', 'Discarder', '-s', '76']
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertEquals(76, parsed.seed)
 
     def test_tournament_simple(self):
         args = ['tournament', 'Discarder',]
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertEquals(2, parsed.per_round)
         self.assertEquals(parsed.command, 'tournament')
         self.assertEquals(1, parsed.iterations)
 
     def test_tournament_per_round(self):
         args = ['tournament', 'Discarder', '-p', '4']
-        parsed = playgame.parse_args(args)
+        parsed = play_game.parse_args(args)
         self.assertEquals(parsed.command, 'tournament')
         self.assertEquals(4, parsed.per_round)
 
@@ -76,7 +76,7 @@ class PlayGameTests(unittest.TestCase):
                     'hanabi_ai.players.example.missing.Missing', #missing player
                     'MockBadPlayer'] #player that does not implement HanabiPlayer
         prepped_players = ['hanabi_ai.players.example_discarder.Discarder']
-        self.assertEqual(prepped_players, playgame.validate_players(players))
+        self.assertEqual(prepped_players, play_game.validate_players(players))
 
     def test_disqualify_player_scored(self):
         scores = {
@@ -86,7 +86,7 @@ class PlayGameTests(unittest.TestCase):
             'E' : 20
         }
         disqualified = Set()
-        playgame.disqualify_player(disqualified, scores, 'D')
+        play_game.disqualify_player(disqualified, scores, 'D')
         self.assertTrue('D' in disqualified)
         self.assertEqual(1, len(disqualified))
         self.assertTrue('D' not in scores)
@@ -99,7 +99,7 @@ class PlayGameTests(unittest.TestCase):
             'E' : 20
         }
         disqualified = Set()
-        playgame.disqualify_player(disqualified, scores, 'C')
+        play_game.disqualify_player(disqualified, scores, 'C')
         self.assertTrue('C' in disqualified)
         self.assertEqual(1, len(disqualified))
         self.assertTrue('C' not in scores)
@@ -112,8 +112,8 @@ class PlayGameTests(unittest.TestCase):
             'E' : 20
         }
         disqualified = Set()
-        playgame.disqualify_player(disqualified, scores, 'A')
-        playgame.disqualify_player(disqualified, scores, 'E')
+        play_game.disqualify_player(disqualified, scores, 'A')
+        play_game.disqualify_player(disqualified, scores, 'E')
         self.assertTrue('E', 'A' in disqualified)
         self.assertEqual(2, len(disqualified))
         self.assertTrue('A', 'E' not in scores)
@@ -127,7 +127,7 @@ class PlayGameTests(unittest.TestCase):
             'E' : {'mean': 25, 'variance': 1},
             'F' : {'mean': 1, 'variance': 1}
         }
-        winners = playgame.determine_winner(results)
+        winners = play_game.determine_winner(results)
         self.assertTrue('A' not in winners)
         self.assertTrue('B' not in winners)
         self.assertTrue('C' not in winners)
@@ -143,7 +143,7 @@ class PlayGameTests(unittest.TestCase):
             'D' : {'mean': 25, 'variance': 4},
             'E' : {'mean': 25, 'variance': 4}
         }
-        winners = playgame.determine_winner(results)
+        winners = play_game.determine_winner(results)
         self.assertTrue('A' not in winners)
         self.assertTrue('B' not in winners)
         self.assertTrue('C' not in winners)
@@ -151,5 +151,5 @@ class PlayGameTests(unittest.TestCase):
         self.assertTrue('E' in winners)
 
     def test_prep_players(self):
-        init_players = playgame.prep_players(['hanabi_ai.players.example_discarder.Discarder'])
+        init_players = play_game.prep_players(['hanabi_ai.players.example_discarder.Discarder'])
         self.assertTrue(isinstance(init_players[0], HanabiPlayer))
